Use relative coordinates for bubbles

+++ schema.sql
@@ -3,8 +3,8 @@ CREATE TABLE IF NOT EXISTS bulles (
   id SERIAL PRIMARY KEY,
   etage TEXT NOT NULL,
   chambre TEXT NOT NULL,
-  x INTEGER,
-  y INTEGER,
+  x REAL,
+  y REAL,
   numero INTEGER,
   description TEXT,
   photo TEXT
